requirements.txt
face-recognition==1.3.0
opencv-python==4.7.0.72
Flask==2.3.2
Werkzeug==2.3.4
numpy==1.24.3
Pillow==9.5.0
python-dotenv==1.0.0 